Get started analyzer code fix (#6824)
* bring in the original files from the roslyn wiki

No edits yet.

* stash - switching branches.

* sample updated

* add notes

* final proofread and ready for review

* fix build errors.

* fix links to source code.

* final set of build warnings.

* major rewrite

This has some significant re-writes based on the earlier reviews. It should flow better and keep the user focused on the steps to update the analyzer.

Tests are used as good examples of code that should or should not raise the diagnostic.

* grammar check and proofread

* fix build warnings.

A few mis-placed links.

* respond to feedback

I've updated up to the first run with the code fix.

* respond to feedback

Thanks for the thorough review @Thraka  All the changes have been addressed.