Added documentation for PackageDescription (#8579)
* Added documentation for PackageDescription

Added documentation for PackageDescription. Adjusted documentation for Description to convey its dual purpose and when it is used. #8578

* Updated Description 

Based upon PR feedback, copied the Description attribute text from main page and modified it to fit the wording of the current page.

* Description change

Adjusted text to clarify behavior of Description when PackageDescription is missing.