Update RIDs for uniformity, remove unsupported OS (#8710)
* Update RIDs for uniformity, remove unsupported OS

Updated the RID list to provide a link to the larger json document in each section for those who jump around and may not read the paragraph leading into the section (this is backed by issues/comments from others who read this as an exhaustive list). Removed Ubuntu 17.10 as it's no longer supported. Reformatted the macOS section to be more consistent with the others by listing each OS version separate and included a "portable" target. Added a note to each "portable" entry to denote it's usable with .NET Core 2.0 or above.

* Updates to document based on review from @jkotas on Pull Request #8710

* Simplified Linux RIDs based on review comments from @jkotas in pull request #8710

* Changed comment on RHEL 6 from alias to superseded per @jkotas review

* update metadata