Fix broken link and provide HttpClientFactory link (#12147)
Fix the link to ResilientHttpClient which has been deprecated. The new link is direct to the source file of the old implementation in the commit which removed it allowing for context on what was added to replace it.

Add link to HttpClientFactory documentation to explain that concept for people first encountering it here.