BOM required to override computer codepage (#11995)
* BOM required to override computer codepage

This page said that **-codepage** is not needed for source code files created with UNICODE or UTF-8. If that's true, then if I have a .cs file encoded in UTF-8 with no Byte Order Mark, csc.exe will interpret it as UTF-8 regardless of the codepage in effect on the computer. That doesn't seem possible, however, since csc.exe can't know whether the file is encoded as UTF-8 or the computer's codepage.

This change corrects the text to indicate that if the source file indicates that uses a Unicode encoding by including a byte order mark, then the computer's codepage is ignored.

* Advise based on initial attempt to parse as UTF-8

Updated advice based on @jaredpar's description of the compiler's parsing heuristics.

* Spelling fix

Co-Authored-By: Bill Wagner <wiwagn@microsoft.com>